+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Pro Formatter Change Log
 
+## Version 0.1.2: December 14, 2021
+
+- **Bugfix**: Fixed file path bug
+- **Bugfix**: Fixed formatting bug
+- **Info**: Removed *showFormatting* due to vscode API issues
+
 ## Version 0.1.1: December 14, 2021
 
 - **Bugfix**: Fixed extension settings bug